restructuring data handling methods
