+++ src/qicklab/analysis/starkspec.py
@@ -2,9 +2,8 @@ import os, datetime
 import numpy as np
 
 ## QICKLAB methods
-from ..datahandling.datafile_tools import load_h5_data
+from ..datahandling.datafile_tools import load_h5_data, process_h5_data
 from ..utils.ana_utils  import rotate_and_threshold
-from ..utils.data_utils import process_h5_data
 from .plot_tools import plot_stark_simple
 from .shot_tools import process_shots
 from .stark_tools import gain2freq_Duffing
