+++ src/qicklab/analysis/t1.py
@@ -4,9 +4,8 @@ import numpy as np
 from scipy.optimize import curve_fit
 
 ## QICKLAB methods
-from ..datahandling.datafile_tools import load_h5_data
+from ..datahandling.datafile_tools import load_h5_data, process_h5_data
 from ..utils.ana_utils  import rotate_and_threshold
-from ..utils.data_utils import process_h5_data
 from .fit_functions import exponential
 from .fit_tools import fit_t1
 from .plot_tools import plot_shots, plot_t1_simple
