+++ src/qicklab/datahandling/datafile_tools.py
@@ -2,8 +2,8 @@
 import os, h5py
 import numpy as np
 
-from ..utils.file_utils import create_h5_dataset
 from ..utils.data_utils import unwrap_singleton_list
+from ..utils.file_utils import create_h5_dataset
 
 
 DATETIME_FMT = "%Y-%m-%d_%H-%M-%S"
@@ -47,103 +47,10 @@ def save_to_h5(data, outer_folder_expt, data_type, batch_num, save_r):
             for key, value in qubit_data.items():
                 create_h5_dataset(key, value, group)
 
-# def load_from_h5(filename, data_type, save_r=1):
-#     """
-#     Load data from an HDF5 file and replicate each dataset value a specified number of times.
-
-#     The function reads each group (assumed to represent a qubit) and for every dataset within,
-#     it replicates the read value save_r times in a list.
-
-#     Parameters:
-#         filename (str): Path to the HDF5 file.
-#         data_type (str): A key to wrap the loaded data under.
-#         save_r (int, optional): Number of times to replicate each loaded dataset. Defaults to 1.
-
-#     Returns:
-#         dict: Nested dictionary with structure:
-#               { data_type: { qubit_index: { dataset_name: [data]*save_r, ... }, ... } }
-#     """
-#     data = {data_type: {}}
-#     with h5py.File(filename, 'r') as f:
-#         for qubit_group in f.keys():
-#             # Convert group name (e.g., "Q1") to a zero-based index.
-#             qubit_index = int(qubit_group[1:]) - 1
-#             qubit_data = {}
-#             group = f[qubit_group]
-#             for dataset_name in group.keys():
-#                 # Read the dataset value and replicate it save_r times.
-#                 replicated = [group[dataset_name][()]] * save_r
-#                 qubit_data[dataset_name] = unwrap_singleton_list(replicated)
-#             data[data_type][qubit_index] = qubit_data
-#     return data
-
-
-# def load_from_h5_with_shotdata(filename, data_type, save_r=1):  # Added save_r as parameter.
-
-#     data = {data_type: {}}  # Initialize the main dictionary with the data_type.
-
-#     with h5py.File(filename, 'r') as f:
-#         for qubit_group in f.keys():
-#             qubit_index = int(qubit_group[1:]) - 1
-#             qubit_data = {}
-#             group = f[qubit_group]
-
-#             for dataset_name in group.keys():
-#                 # Attempt to map HDF5 keys to the target dictionaries' keys.
-#                 if data_type == 'Res':
-#                     target_keys = {'Dates': 'Dates', 'freq_pts': 'freq_pts', 'freq_center': 'freq_center',
-#                                        'Amps': 'Amps', 'Found Freqs': 'Found Freqs', 'Round Num': 'Round Num',
-#                                        'Batch Num': 'Batch Num','Exp Config': 'Exp Config', 'Syst Config': 'Syst Config'}
-#                 elif data_type == 'QSpec':
-#                     target_keys = {'Dates': 'Dates', 'I': 'I', 'Q': 'Q', 'Frequencies': 'Frequencies',
-#                                        'I Fit': 'I Fit', 'Q Fit': 'Q Fit', 'Round Num': 'Round Num',
-#                                        'Batch Num': 'Batch Num','Exp Config': 'Exp Config', 'Syst Config': 'Syst Config'}
-#                 elif data_type == 'Ext_QSpec':
-#                     target_keys = {'Dates': 'Dates', 'I': 'I', 'Q': 'Q', 'Frequencies': 'Frequencies','Round Num': 'Round Num',
-#                                        'Batch Num': 'Batch Num', 'Exp Config': 'Exp Config', 'Syst Config': 'Syst Config'}
-
-#                 elif data_type == 'Rabi':
-#                     target_keys = {'Dates': 'Dates', 'I': 'I', 'Q': 'Q', 'Gains': 'Gains', 'Fit': 'Fit',
-#                                        'Round Num': 'Round Num', 'Batch Num': 'Batch Num','Exp Config': 'Exp Config', 'Syst Config': 'Syst Config'}
-#                 elif data_type == 'SS':
-#                     target_keys = {'Fidelity': 'Fidelity', 'Angle': 'Angle', 'Dates': 'Dates', 'I_g': 'I_g',
-#                                        'Q_g': 'Q_g', 'I_e': 'I_e', 'Q_e': 'Q_e',
-#                                        'Round Num': 'Round Num', 'Batch Num': 'Batch Num','Exp Config': 'Exp Config', 'Syst Config': 'Syst Config'}
-#                 elif data_type == 'T1':
-#                     target_keys = {'T1': 'T1', 'Errors': 'Errors', 'Dates': 'Dates', 'I': 'I', 'Q': 'Q',
-#                                        'Delay Times': 'Delay Times', 'Fit': 'Fit', 'Round Num': 'Round Num',
-#                                        'Batch Num': 'Batch Num','Exp Config': 'Exp Config', 'Syst Config': 'Syst Config'}
-#                 elif data_type == 'T2':
-#                     target_keys = {'T2': 'T2', 'Errors': 'Errors', 'Dates': 'Dates', 'I': 'I', 'Q': 'Q',
-#                                        'Delay Times': 'Delay Times', 'Fit': 'Fit', 'Round Num': 'Round Num',
-#                                        'Batch Num': 'Batch Num','Exp Config': 'Exp Config', 'Syst Config': 'Syst Config'}
-#                 elif data_type == 'T2E':
-#                     target_keys = {'T2E': 'T2E', 'Errors': 'Errors', 'Dates': 'Dates', 'I': 'I', 'Q': 'Q',
-#                                        'Delay Times': 'Delay Times', 'Fit': 'Fit', 'Round Num': 'Round Num',
-#                                        'Batch Num': 'Batch Num','Exp Config': 'Exp Config', 'Syst Config': 'Syst Config'}
-#                 elif data_type == 'stark2D':
-#                     target_keys = {'Dates': 'Dates', 'I':'I', 'Q': 'Q', 'Qu Frequency Sweep':'Qu Frequency Sweep',
-#                                    'Res Gain Sweep':'Res Gain Sweep','Round Num':'Round Num', 'Batch Num': 'Batch Num',
-#                                    'Exp Config': 'Exp Config', 'Syst Config': 'Syst Config'}
-#                 elif data_type =='starkSpec':
-#                     target_keys = {'Dates': 'Dates', 'I':'I', 'Q': 'Q','P': 'P', 'shots':'shots','Gain Sweep':'Gain Sweep','Round Num':'Round Num', 'Batch Num': 'Batch Num',
-#                                    'Exp Config': 'Exp Config', 'Syst Config': 'Syst Config'}
-
-#                 else:
-#                     raise ValueError(f"Unsupported data_type: {data_type}")
-
-#                 try:
-#                     mapped_key = target_keys[dataset_name]  # Map HDF5 key to target key.
-#                     qubit_data[mapped_key] = [group[dataset_name][()]] * save_r  # Expand to match the desired length.
-
-#                 except KeyError:
-#                     print(f"Warning: Key '{dataset_name}' not found in target dictionary for data_type '{data_type}'. Skipping.")
-#                     pass
-
-#             data[data_type][qubit_index] = qubit_data
-
-#     return data
-
+def find_h5_files(basepath, dataset, expt_name, folder="study_data"):
+    data_path = os.path.join(basepath, dataset, folder, "Data_h5", expt_name)
+    h5_files = os.listdir(data_path).sort()
+    return h5_files, len(h5_files)
 
 def load_h5_data(filename, data_type, save_r=1):
     """
@@ -210,4 +117,50 @@ def load_h5_data(filename, data_type, save_r=1):
             data[data_type][qubit_index] = qubit_data
 
     return data
+
+def process_h5_data(data):
+    """
+    Process a string containing numeric data by filtering out unwanted characters.
+
+    The function removes newline characters and keeps only digits, minus signs, dots, spaces, and the letter 'e'
+    (for exponential notation), then converts the resulting tokens to floats.
+
+    Parameters:
+        data (str): String containing numeric data.
+
+    Returns:
+        numbers: A list of floats extracted from the string.
+    """
+    # Check if the data is a byte string; decode if necessary.
+    if isinstance(data, bytes):
+        data_str = data.decode()
+    elif isinstance(data, str):
+        data_str = data
+    else:
+        raise ValueError("Unsupported data type. Data should be bytes or string.")
+
+    data_str = data_str.strip().replace('\n', ' ')
+
+    # Remove extra whitespace and non-numeric characters.
+    cleaned_data = ''.join(c for c in data_str if c.isdigit() or c.lower() in ['+', '-', '.', ' ', 'e'])
+
+    # Split into individual numbers, removing empty strings.
+    numbers = [float(x) for x in cleaned_data.split() if x]
+    return numbers
+
+
+def get_data_field(data_dict, expt_name, qubit_idx, data_field, steps=None, reps=None):
+    data = process_h5_data(data_dict[expt_name][int(qubit_idx)].get(data_field, [])[0][0].decode())
+    if (reps is not None) and (steps is not None):
+        return np.array(data).reshape([steps, reps])
+    else:
+        return np.array(data)
+
+# def get_gainsweep(data_dict, expt_name, qubit_idx, data_field, steps=None, reps=None):
+
+
+
+
+
+
